Update .gitignore, add JUnit dependency, and implement TOTP tests
